backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
google-generativeai
requests
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
email-validator